+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - Improved interrupt handling allowing Ctrl+C or closing worker windows to abort without waiting for timeouts (Thank you to "UncleJ4ck")
  - Added band data to Evil twin and WEP control windows, WPS attacks and different trophy files
  - Fixed bug in WPS wash scan target list that skipped first detected networks
+ - Improved WPS wash autoselection handling for single locked targets
 
 ### 12.0
  - Fixed enterprise certificate capture results persisting when switching targets